Updated TC field types in build structures according to application.wadl.

+++ DataSource.TC/DataClasses/builds.cs
@@ -13,7 +13,7 @@ namespace DataSource.TC.DataClasses
 
 		/// <remarks />
 		[XmlAttribute]
-		public byte count { get; set; }
+		public int count { get; set; }
 
 		/// <remarks />
 		[XmlAttribute]
